AutoGluonMultimodal/requirements.txt
autogluon.multimodal>=1.0.0
mlflow>=2.8.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
